requirements.txt
# CIRIS Agent Requirements - Runtime Dependencies Only
# Last updated: 2025-01-08
# Python >= 3.12 required (for PEP 695 type parameter syntax)

# Core Dependencies
pydantic>=2.7.0,<3.0.0
openai>=1.0.0,<2.0.0
anthropic>=0.40.0,<1.0.0  # Native Anthropic SDK for Claude models
google-generativeai>=0.8.0,<1.0.0  # Native Google SDK for Gemini models (legacy)
google-genai>=1.0.0,<2.0.0  # New Google GenAI SDK with instructor support
jsonref>=1.0.0,<2.0.0  # Required by google-genai for schema resolution
instructor>=1.11.3,<2.0.0
PyYAML>=6.0.1,<7.0.0

# Database
psycopg2-binary>=2.9.0,<3.0.0  # PostgreSQL adapter for production deployments
SQLAlchemy>=2.0.0,<3.0.0  # SQL toolkit and ORM for external data sources

# Async and HTTP
aiohttp>=3.9.0,<4.0.0
httpx>=0.25.0,<1.0.0
anyio>=4.0.0,<5.0.0

# Discord Integration
discord.py>=2.3.0,<3.0.0

# Document Processing (Security-First)
pypdf>=6.0.0,<7.0.0  # PDF text extraction - CVE-2024-XXXXX (FlateDecode RAM exhaustion) fixed in 6.0.0
docx2txt>=0.8,<1.0.0  # Pure Python DOCX text extraction

# Security and Cryptography
cryptography>=41.0.0,<46.0.0
PyJWT[crypto]>=2.8.0,<3.0.0
bcrypt>=4.0.0,<5.0.0
PyNaCl>=1.5.0,<2.0.0
ciris-verify>=1.0.8  # Hardware-rooted license verification (CIRISVerify FFI bindings)

# Web Framework (API)
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
websockets>=14.0
python-multipart>=0.0.6,<1.0.0
aiofiles>=23.2.0,<24.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
networkx>=3.0,<4.0
psutil>=5.9.0,<6.0.0
croniter>=2.0.0,<3.0.0
backoff>=2.2.0,<3.0.0

# Network Discovery
zeroconf>=0.39.0,<1.0.0  # mDNS/DNS-SD discovery for Home Assistant (0.39.x for Python 3.10, 0.80+ for Python 3.11+)

# CLI
click>=8.1.0,<9.0.0

# Container Management
docker>=6.0.0,<8.0.0
